+++ tests/testthat/test-bayesfactor_models.R
@@ -36,14 +36,14 @@ test_that("bayesfactor_models RSTANARM", {
   testthat::skip("Skipping bayesfactor_models RSTANARM")
   library(rstanarm)
   set.seed(444)
-  junk <- capture.output(stan_bf_0 <- stan_glm(Sepal.Length ~ 1,
-    data = iris,
-    diagnostic_file = file.path(tempdir(), "df0.csv")
-  ))
-  junk <- capture.output(stan_bf_1 <- stan_glm(Sepal.Length ~ Species,
-    data = iris,
-    diagnostic_file = file.path(tempdir(), "df1.csv")
-  ))
+  stan_bf_0 <- stan_glm(Sepal.Length ~ 1,
+                        data = iris,
+                        refresh = 0,
+                        diagnostic_file = file.path(tempdir(), "df0.csv"))
+  stan_bf_1 <- stan_glm(Sepal.Length ~ Species,
+                        data = iris,
+                        refresh = 0,
+                        diagnostic_file = file.path(tempdir(), "df1.csv"))
 
   testthat::expect_warning(bayestestR::bayesfactor_models(stan_bf_0, stan_bf_1))
   stan_models <- suppressWarnings(bayestestR::bayesfactor_models(stan_bf_0, stan_bf_1))