+++ tests/testthat/test-bayesfactor_parameters.R
@@ -46,14 +46,14 @@ test_that("bayesfactor_parameters RSTANARM", {
 
   library(rstanarm)
   set.seed(333)
-  junk <- capture.output(model <- stan_glm(extra ~ group, data = sleep))
+  model <- stan_glm(extra ~ group, data = sleep, refresh = 0)
   bfsd <- bayestestR::bayesfactor_parameters(model)
   testthat::expect_equal(log(bfsd$BF), c(-2.69, -0.14), tolerance = 0.2)
 
   bfsd <- bayestestR::bayesfactor_parameters(model, null = rope_range(model))
   testthat::expect_equal(log(bfsd$BF), c(-2.96, -0.18), tolerance = 0.2)
 
-  junk <- capture.output(model_p <- update(model, prior_PD = TRUE))
+  model_p <- update(model, prior_PD = TRUE, refresh = 0)
   bfsd <- bayestestR::bayesfactor_parameters(model, model_p)
   testthat::expect_equal(log(bfsd$BF), c(-2.69, -0.14), tolerance = 0.2)
 })